+++ RoadAPI/Services/ReportRepository.cs
@@ -101,7 +101,6 @@ namespace RoadAPI.Services
                     item.Content = model.Content;
                 }
 
-                _mapper.Map(model, item);
                 if (model.Image != null)
                 {
                     deleteImage(item.Image);
@@ -171,16 +170,19 @@ namespace RoadAPI.Services
             }
         }
 
-        private void deleteImage(string fileName)
+        private void deleteImage(string? fileName)
         {
-            path = Path.Combine(Directory.GetCurrentDirectory(), path);
-
-            // Get the path to the image file
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), path, fileName);
-
-            if (File.Exists(filePath))
+            if (fileName != null)
             {
-                File.Delete(filePath);
+                path = Path.Combine(Directory.GetCurrentDirectory(), path);
+
+                // Get the path to the image file
+                var filePath = Path.Combine(Directory.GetCurrentDirectory(), path, fileName);
+
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath);
+                }
             }
         }
     }
